refactor: RandomNumberGenerator 간결하게 변경
