+++ src/main/java/baseball/core/RandomNumberGenerator.java
@@ -1,21 +1,21 @@
-package baseball.core.opponent;
+package baseball.core;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class RandomNumberGenerator {
 
     private static final int START_INCLUSIVE = 1;
     private static final int END_INCLUSIVE = 9;
-    private static final int COUNT = 3;
+    private static final int MAX_BALL_COUNT = 3;
 
     public static List<Integer> generate() {
-        Stream.generate(() -> Randoms.pickNumberInRange(START_INCLUSIVE, END_INCLUSIVE))
+        return Stream.generate(() -> Randoms.pickNumberInRange(START_INCLUSIVE, END_INCLUSIVE))
                 .distinct()
-                .limit(COUNT)
-        return threeNumbers;
+                .limit(MAX_BALL_COUNT)
+                .collect(Collectors.toList());
     }
 }
